Multiple fixes for building with gcc
- Updated genmake
- Linux path for .project file and in build.sh
- Fixed compile errors

+++ build.sh
@@ -12,6 +12,19 @@ Tools/ucc --type=typeinfo.bin Core Anim Editor || exit
 [ "$OSTYPE" == "linux-gnu" ] || [ "$OSTYPE" == "linux" ] && PLATFORM="linux"
 #[ "$PLATFORM" == "linux" ] && PLATFORM="linux64"
 
+if [ "${PLATFORM:0:3}" == "vc-" ]; then
+	# Visual C++ compiler
+	# setup default compiler version
+	[ "$vc_ver" ] || vc_ver=latest
+	# Find Visual Studio
+	. vc32tools $VC32TOOLS_OPTIONS --version=$vc_ver --check
+	[ -z "$found_vc_year" ] && exit 1				# nothing found
+	# Adjust vc_ver to found one
+	vc_ver=$found_vc_year
+#	echo "Found: $found_vc_year $workpath [$vc_ver]"
+	GENMAKE_OPTIONS=VC_VER=$vc_ver					# specify compiler for genmake script
+fi
+
 # build resources archive
 rm -f $resfile
 pkzipc -add $resfile -lev=9 -nozip -silent xrc/*.png xrc/*.xrc
@@ -21,7 +34,21 @@ makefile="obj/SkelEdit-$PLATFORM.mak"
 if ! [ -d obj ]; then
 	mkdir obj
 fi
-Tools/genmake $project TARGET=$PLATFORM > $makefile
+Tools/genmake $project TARGET=$PLATFORM $GENMAKE_OPTIONS > $makefile
 
 # build
-vc32tools --version=2015 --make $makefile
+# if $target is empty, whole project will be built, otherwise a single file
+case "$PLATFORM" in
+	"vc-win32")
+		Make $makefile $target || exit 1
+		;;
+	"vc-win64")
+		Make $makefile $target || exit 1
+		;;
+	linux*)
+		make -j 4 -f $makefile $target || exit 1	# use 4 jobs for build
+		;;
+	*)
+		echo "Unknown PLATFORM=\"$PLATFORM\""
+		exit 1
+esac
